fix: correção das entidades

+++ src/main/java/dev/luanc/library/repository/LoanRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface LoanRepository extends JpaRepository<Loan, Long> {
+public interface LoanRepository extends JpaRepository<Loan, Integer> {
     public List<Loan> findLoanByUserEmail(String email);
 }
